+++ CHANGELOG.md
@@ -13,9 +13,16 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## v3.0.0-rc.??? (2021-??-??)
 
+#### :rocket: New Feature
+
+* Added a new option `--bail` to stop test execution after the first failure `tests`
+
 #### :bug: Bug Fix
 
 * [Fixed an issue with tests execution took too long](https://github.com/V4Fire/Client/issues/436)
+
+* Added `force=true` to default options `DOM.clickToRef`. It fixed a bug where clicking at a visible element
+  did not work because the playwright decided that the element was not visible. `tests`
 
 #### :house: Internal